menu title is now a MenuComponent for more flexibility
