+++ application/src/main/java/com/agonyforge/mud/cli/demo/MenuQuestion.java
@@ -6,13 +6,15 @@ import com.agonyforge.mud.cli.menu.Color;
 import com.agonyforge.mud.cli.menu.demo.DemoMenuItem;
 import com.agonyforge.mud.cli.menu.demo.DemoMenuPane;
 import com.agonyforge.mud.cli.menu.demo.DemoMenuPrompt;
+import com.agonyforge.mud.cli.menu.demo.DemoMenuTitle;
 import com.agonyforge.mud.web.model.Input;
 import com.agonyforge.mud.web.model.Output;
 
 public class MenuQuestion implements Question {
-    private final DemoMenuPane menuPane = new DemoMenuPane("Demo Menu");
+    private final DemoMenuPane menuPane = new DemoMenuPane();
 
     public MenuQuestion() {
+        menuPane.setTitle(new DemoMenuTitle("Demo Menu"));
         menuPane.getItems().add(new DemoMenuItem("F", "Foo"));
         menuPane.getItems().add(new DemoMenuItem("B", "Bar"));
         menuPane.getItems().add(new DemoMenuItem("C", "Crazy Town"));
