+++ application/src/main/java/com/agonyforge/mud/cli/menu/AbstractMenuPane.java
@@ -6,24 +6,20 @@ import java.util.ArrayList;
 import java.util.List;
 
 public abstract class AbstractMenuPane implements MenuComponent {
-    private String title;
+    private AbstractMenuTitle title;
     private final List<AbstractMenuItem> items = new ArrayList<>();
     private AbstractMenuPrompt prompt;
-
-    public AbstractMenuPane(String title) {
-        setTitle(title);
-    }
 
     public List<AbstractMenuItem> getItems() {
         return items;
     }
 
-    public void setTitle(String title) {
-        this.title = title;
+    public AbstractMenuTitle getTitle() {
+        return title;
     }
 
-    public String getTitle() {
-        return title;
+    public void setTitle(AbstractMenuTitle title) {
+        this.title = title;
     }
 
     public AbstractMenuPrompt getPrompt() {
